requirements.txt
python-dotenv
docker
psutil
selenium
webdriver-manager
